move git commands to common/git.py (#31921)
move git into common/git.py
