+++ common/utils.py
@@ -1,3 +1,11 @@
+from collections.abc import Callable
+from functools import lru_cache
+from typing import TypeVar
+
+
+_RT = TypeVar("_RT")
+
+
 class Freezable:
   _frozen: bool = False
 
@@ -9,3 +17,7 @@ class Freezable:
     if self._frozen:
       raise Exception("cannot modify frozen object")
     super().__setattr__(*args, **kwargs)
+
+
+def cache(user_function: Callable[..., _RT], /) -> Callable[..., _RT]:
+  return lru_cache(maxsize=None)(user_function)
