+++ selfdrive/manager/manager.py
@@ -16,9 +16,10 @@ from openpilot.selfdrive.manager.process import ensure_running
 from openpilot.selfdrive.manager.process_config import managed_processes
 from openpilot.selfdrive.athena.registration import register, UNREGISTERED_DONGLE_ID
 from openpilot.common.swaglog import cloudlog, add_file_handler
-from openpilot.system.version import is_dirty, get_commit, get_version, get_origin, get_short_branch, \
+from openpilot.common.git import get_commit, get_origin, get_short_branch, get_commit_date
+from openpilot.system.version import is_dirty, get_version, \
                            get_normalized_origin, terms_version, training_version, \
-                           is_tested_branch, is_release_branch, get_commit_date
+                           is_tested_branch, is_release_branch
