+++ selfdrive/test/process_replay/test_debayer.py
@@ -9,7 +9,7 @@ import pyopencl as cl  # install with `PYOPENCL_CL_PRETEND_VERSION=2.0 pip insta
 from openpilot.system.hardware import PC, TICI
 from openpilot.common.basedir import BASEDIR
 from openpilot.tools.lib.openpilotci import BASE_URL
-from openpilot.system.version import get_commit
+from openpilot.common.git import get_commit
 from openpilot.system.camerad.snapshot.snapshot import yuv_to_rgb
 from openpilot.tools.lib.logreader import LogReader
 from openpilot.tools.lib.filereader import FileReader
